Home Accessory Architect v10.1.1 Kestrel

+++ devices/HAA/setup_mode/src/wifi_config.c
@@ -165,9 +165,9 @@ static void client_send(client_t *client, const char *payload, size_t payload_si
 }
 
 static void client_send_chunk(client_t *client, const char *payload) {
-    int len = strlen(payload);
+    size_t len = strlen(payload);
     char buffer[10];
-    int buffer_len = snprintf(buffer, sizeof(buffer), "%x\r\n", len);
+    size_t buffer_len = snprintf(buffer, sizeof(buffer), "%x\r\n", len);
     client_send(client, buffer, buffer_len);
     client_send(client, payload, len);
     client_send(client, "\r\n", 2);
